+++ src/components/vulnerability-db/VulDBDashboard.tsx
@@ -58,15 +58,15 @@ function VulDBListCard({ data }: { data: VulDBPost }) {
 
           {data.source === "CNNVD" &&
             isCnnvdContentType(data.content) &&
-            (Array.isArray(data.content?.description?.original) ? (
-              data.content.description.original.map((item, index) => (
+            (Array.isArray(data.content?.description?.translated) ? (
+              data.content.description.translated.map((item, index) => (
                 <span key={index} className="text-[#797979]">
                   {item || "No Description Available"}
                 </span>
               ))
             ) : (
               <span className="text-[#797979]">
-                {data.content?.description?.original ||
+                {data.content?.description?.translated ||
                   "No Description Available"}
               </span>
             ))}
